Divide commands to their separate files.
All the commands are now located in their corresponding files
under the cmds directory.
